DDF-5900 update for security api and impl split
