+++ pom.xml
@@ -51,7 +51,7 @@
     <maven.deploy.plugin.version>2.8.2</maven.deploy.plugin.version>
 
     <!--Backend properties-->
-    <ddf.version>2.23.2</ddf.version>
+    <ddf.version>2.24.0-SNAPSHOT</ddf.version>
     <ddf.support.version>2.3.16</ddf.support.version>
     <karaf.version>4.2.6</karaf.version>
     <camel.version>2.22.1</camel.version>
