+++ ui-backend/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/metacard/MetacardApplication.java
@@ -80,8 +80,8 @@ import ddf.catalog.source.UnsupportedQueryException;
 import ddf.catalog.transform.QueryResponseTransformer;
 import ddf.catalog.util.impl.ResultIterable;
 import ddf.security.SubjectIdentity;
-import ddf.security.SubjectUtils;
 import ddf.security.common.audit.SecurityLogger;
+import ddf.security.impl.SubjectUtils;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
